+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update client tests for fetch-based SSE and QueryClient mock
 ---
 
 ## [0.26.0] - 2026-03-28